+++ ChangeLog.md
@@ -1,5 +1,10 @@
 # JPSUN - Notes de versions (ChangeLog)
 
+## 1.8 (18/01/2026)
+
+- Création automatique d'un projet validé à la signature d'un devis, avec liens aux objets et copie des extrachamps.
+- Ajout d'un réglage pour activer la création automatique de projet et mise à jour du libellé projet avec la référence client.
+
 ## 1.7 (11/01/2026)
 
 - Ajout des extrafields contrats JPSUN pour les contrats.